backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
google-generativeai==0.7.2
pydantic==2.11.10
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
httpx==0.28.1
pytest==8.3.5
numpy>=1.26.0
